+++ src/Symfony/Component/Serializer/Normalizer/NormalizedValueInterface.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Symfony\Component\Serializer\Normalizer;
+
+interface NormalizedValueInterface {
+
+    public function getNormalization(): array|string|int|float|bool|\ArrayObject|null;
+    
+}